\name{AwardsMisc}
\alias{AwardsMisc}
\docType{data}
\title{
AwardsMisc Table
}
\description{
Award information for miscellaneous awards such as the Lester Patrick and
general manager of the year awards.
}
\usage{data(AwardsMisc)}
\format{
  A data frame with observations on the following 6 variables.
  \describe{
    \item{\code{name}}{Full name of award recipient}
    \item{\code{ID}}{Recipient ID code}
    \item{\code{award}}{Name of award won}
    \item{\code{year}}{Year (2005-06 is listed as \code{2005})}
    \item{\code{lgID}}{League ID}
    \item{\code{note}}{Notes about the award}
  }
}
\source{
Hockey Databank: \url{http://groups.yahoo.com/neo/groups/hockey-databank/info}
}
\examples{
data(AwardsMisc)
}
\keyword{datasets}
